+++ src/FilterHandler/AnyHandler.php
@@ -6,7 +6,7 @@ namespace Yiisoft\Data\Db\FilterHandler;
 
 use Yiisoft\Data\Reader\Filter\Any;
 
-final class AnyHandler extends BaseHandler
+final class AnyHandler extends GroupHandler
 {
     public function getOperator(): string
     {
