+++ src/FilterHandler/GreaterThanHandler.php
@@ -6,7 +6,7 @@ namespace Yiisoft\Data\Db\FilterHandler;
 
 use Yiisoft\Data\Reader\Filter\GreaterThan;
 
-final class GreaterThanHandler extends BaseHandler
+final class GreaterThanHandler extends CompareHandler
 {
     public function getOperator(): string
     {
