+++ src/FilterHandler/LessThanHandler.php
@@ -6,7 +6,7 @@ namespace Yiisoft\Data\Db\FilterHandler;
 
 use Yiisoft\Data\Reader\Filter\LessThan;
 
-final class LessThanHandler extends BaseHandler
+final class LessThanHandler extends CompareHandler
 {
     public function getOperator(): string
     {
